% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata_xml.R
\name{create_metadata_xml}
\alias{create_metadata_xml}
\title{Create xml for i2b2 ontology in C_METADATAXML}
\source{
https://community.i2b2.org/wiki/display/DevForum/Metadata+XML+for+Medication+Modifiers
}
\usage{
create_metadata_xml(
  version = "3.02",
  creation_datetime,
  test_id,
  test_name = test_id,
  data_type,
  flags_to_use,
  ok_to_use_values = "Y",
  unit_values = list(),
  enum_values = list(),
  max_string_length = NULL
)
}
\arguments{
\item{version}{Metadataxml version (default: 3.02)}

\item{creation_datetime}{Creation datetime in format 01/26/2011 00:00:00}

\item{test_id}{Code for the value}

\item{test_name}{Came of the value}

\item{data_type}{Data type of value.  Allowed data_types are: PosInteger,
Integer, PosFloat, Float, Enum, or String}

\item{flags_to_use}{Allowed valueflag_cds.  Allowed flags N, L, H, A, T.
Multiple flags can be specified in a single string (ex: NLH). Set to blank
string for no flag constraints.}

\item{ok_to_use_values}{default to Y.  Any other value will disable value
constraints}

\item{unit_values}{A list of allowable unit values}

\item{enum_values}{A list of allowable enum values with description}

\item{max_string_length}{Max character length for string searches.}
}
\value{
XML object
}
\description{
Create xml for i2b2 ontology in C_METADATAXML
}
\examples{
unit_values <- list(
 NormalUnits = "mg/dose",
 EqualUnits = c("mg/tablet", "gm/liter"),
 ConvertingUnits = list(
  list(Units = "gm/tablet", MultiplyingFactor = 1000),
  list(Units = "mg/0.5ml", MultiplyingFactor = 2)),
 ExcludingUnits = c("\%", "iu"))

 enum_values <- list(Val = list(
  list(description = "Before meals", value = "AC"),
  list(description = "Twice per day", value = "BID")),
 ExcludingVal = list(list(
  description = "Invalid result", value = "pending")))

 xml_output <- create_metadata_xml(
  version = "3.02",
  creation_datetime = "01/26/2011 00:00:00",
  test_id = "MED:DOSE",
  test_name = "Medication Dose",
  data_type = "PosFloat",
  flags_to_use = "NHL",
  ok_to_use_values = "Y",
  unit_values = unit_values,
  enum_values = enum_values
  )

  cat(xml_output)
}
