% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_adult_enrollment_example.R
\docType{data}
\name{recover_adult_enrollment_data}
\alias{recover_adult_enrollment_data}
\title{RECOVER Adult Enrollment Data}
\format{
## `recover_adult_enrollment_data`
A data frame with 211 rows and 39 columns:
\describe{
  \item{record_id}{RECOVER participant ID}
  \item{redcap_event_name}{The REDCap event name.  Enrollment data will only
  exist in the enrollment_arm_1 event but the dataset includes other events
  for testing and development purposes.}
  \item{redcap_repeat_instrument}{Form name for repeating instances}
  \item{redcap_repeat_instance}{Number of repeat instance}
  \item{redcap_data_access_group}{DAG of data entry}
  \item{enrollment_fversion - enrollment_complete}{Data from the enrollment
  form.  See data dictionary for field definitions}

}
}
\usage{
recover_adult_enrollment_data
}
\description{
Test data from the RECOVER Adult Observational Study.  The data subset
only includes REDCap fields from the study enrollment REDCap instrument.
This is fictional data and does not include any real participant information.
The data was pulled via REDCap API from the RECOVER REDCap Sandbox on
2024-07-08.
}
\keyword{datasets}
