% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_adult_enrollment_example.R
\docType{data}
\name{recover_adult_enrollment_dd}
\alias{recover_adult_enrollment_dd}
\title{RECOVER Adult Enrollment Data Dictionary}
\format{
## `recover_adult_enrollment_dd`
A data frame with 44 rows and 18 columns:
\describe{
  \item{field_name}{Field name}
  \item{form_name}{Instrument name}
  \item{section_header}{The name of the section}
  \item{field_type}{Type of field (e.g. text, radio, checkbox)}
  \item{field_label}{Text label of field.  May include HTML and other special
  characters supported in REDCap}
  \item{select_choices_or_calculations}{Choices for radio and checkbox fields
  and calculations for calc field_types}
  \item{field_note}{Note for the field}
  \item{text_validation_type_or_show_slider_number}{Validation type for text
  fields (which can include numeric inputs)}
  \item{text_validation_min}{Minimum allowed value in text field}
  \item{text_validation_max}{Maximum allowed value in text field}
  \item{text_validation_max}{logical indication if the field is tagged as
  an identifier}
  \item{branching_logic}{Branching logic in REDCap syntax}
  \item{required_field}{Is the field required}
  \item{custom_alignment}{Position of field on the REDCap form UI}
  \item{question_number}{Number of the question}
  \item{matrix_group_name}{Grouping for fields organized in a matrix}
  \item{matrix_ranking}{Ranking for fields organized in a matrix}
  \item{field_annotation}{Annotations and smart field logic}

}
}
\usage{
recover_adult_enrollment_dd
}
\description{
Data dictionary from the RECOVER Adult Longitudinal Study.  The dictionary
only include REDCap fields from the study enrollment REDCap instrument.
The data was pulled via REDCap API from the RECOVER REDCap Sandbox on
2024-07-08.
}
\keyword{datasets}
