% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_adult_enrollment_example.R
\docType{data}
\name{recover_adult_enrollment_inst}
\alias{recover_adult_enrollment_inst}
\title{RECOVER Adult Enrollment Instrument Labels}
\format{
## `recover_adult_enrollment_dd`
A data frame with 5 rows and 2 columns:
\describe{
  \item{instrument_name}{Instrument name corresponding to form_name in data
  dictionary}
  \item{instrument_label}{Instrument nice label}

}
}
\usage{
recover_adult_enrollment_inst
}
\description{
Instrument labels from the RECOVER Adult Longitudinal Study.  The dictionary
only include a few REDCap instrument mappings.
The data was pulled via REDCap API from the RECOVER REDCap Sandbox on
2024-07-08.
}
\keyword{datasets}
