% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap_data.R
\name{tidy_redcap_data}
\alias{tidy_redcap_data}
\title{Tidy REDCap data}
\usage{
tidy_redcap_data(redcap_data, redcap_dd)
}
\arguments{
\item{redcap_data}{a REDCap data file}

\item{redcap_dd}{a REDCap data dictionary}
}
\value{
A data.frame
}
\description{
Tidy REDCap data
}
