% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redcap_dictionary.R
\name{tidy_redcap_dictionary}
\alias{tidy_redcap_dictionary}
\title{Tidy REDCap Data Dictionary}
\usage{
tidy_redcap_dictionary(
  redcap_dd,
  instrument_labels = NULL,
  redcap_vars = NA,
  section_mappings = NULL,
  calc_field_settings = NULL
)
}
\arguments{
\item{redcap_dd}{REDCap data dictionary}

\item{instrument_labels}{REDCap instrument labels}

\item{redcap_vars}{REDCap vars file}

\item{section_mappings}{Custom section mappings}

\item{calc_field_settings}{Custom calculated field settings}
}
\value{
data.frame
}
\description{
Tidy REDCap Data Dictionary
}
