% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_i2b2_facts.R
\name{redcap_i2b2_facts}
\alias{redcap_i2b2_facts}
\title{Build an i2b2 observation_fact table from a REDCap tidy data and dictionary}
\usage{
redcap_i2b2_facts(
  redcap_data,
  redcap_i2b2_ontology,
  project_id = "REDCap",
  date_mappings
)
}
\arguments{
\item{redcap_data}{REDCap tidy data}

\item{redcap_i2b2_ontology}{i2b2 ontology (usually from output of
redcap_i2b2_ontology())}

\item{project_id}{Project name to use as the root level (default: REDCap)}

\item{date_mappings}{Named list with fields to use for the start_date of
each form}
}
\value{
data.frame
}
\description{
Build an i2b2 observation_fact table from a REDCap tidy data and dictionary
}
