% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_i2b2_ontology.R
\name{redcap_i2b2_ontology}
\alias{redcap_i2b2_ontology}
\title{Build an i2b2 ontology table from a REDCap dictionary}
\usage{
redcap_i2b2_ontology(
  redcap_tidy_dd,
  project_id = "REDCap",
  concept_cd_prefix = "RC:",
  include_forms = NA,
  strip_html = TRUE
)
}
\arguments{
\item{redcap_tidy_dd}{REDCap tidy data dictionary}

\item{project_id}{Project name to use as the root level (default: REDCap)}

\item{concept_cd_prefix}{Prefix to prepend to C_BASECODE (default: RC:)}

\item{include_forms}{Forms to include in the i2b2 ontology.  The default is
to include all forms}

\item{strip_html}{Boolean to strip html from section_header and field_label.
The default is TRUE.}
}
\value{
data.frame
}
\description{
Build an i2b2 ontology table from a REDCap dictionary
}
